requirements.txt
discord.py 
openai-whisper 
elevenlabs 
numpy
discord.py[voice]
discord-ext-voice-recv
torch
aiodns
aiohttp